requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
deep-translator==1.11.4
Faker==18.13.0
google-cloud-translate==3.12.1
setuptools==68.2.2
scapy==2.5.0
pyautogui==0.9.54
webdriver-manager==3.0.0
pytest==8.2.0
python-dateutil==2.9.0
mouseinfo==0.1.3
pillow==9.5.0
pyvirtualdisplay==3.0
python-xlib==0.33
